Dao test with abstract DaoTest
