+++ test/com/codecool/shop/dao/implementation/ProductDaoTest.java
@@ -1,37 +1,13 @@
 package com.codecool.shop.dao.implementation;
 
-import com.codecool.shop.dao.ProductDao;
 import com.codecool.shop.model.Product;
-import com.codecool.shop.model.ProductCategory;
-import com.codecool.shop.model.Supplier;
-
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class ProductDaoTest {
-
-    ProductDao productDao;
-    Product productFirst;
-    Product productSecond;
-    Supplier supplierFirst;
-    Supplier supplierSecond;
-    ProductCategory productCategoryFirst;
-    ProductCategory productCategorySecond;
-
-    @Before
-    public void setUp(){
-        productDao = ProductDaoMem.getInstance();
-        productCategoryFirst = new ProductCategory("test", "oldschool", "first test for dao");
-        productCategorySecond = new ProductCategory("test2", "oldschool2", "second test for dao");
-        supplierFirst = new Supplier("codecool.bp.1", "spring class");
-        supplierSecond = new Supplier("codecool.bp.2", "autumn class");
-        productFirst = new Product("daotest",49.9f, "USD", "Newbie to testing", productCategoryFirst, supplierFirst);
-        productSecond = new Product("daotest2",59.9f, "USD", "Getting start for testing", productCategorySecond, supplierSecond);
-    }
+public class ProductDaoTest extends DaoTest{
 
     @Test
     public void testAdd() throws Exception{
